Module Soteria_c_lib.Config

type t = {

auto_include_path : string option;
(*
Path to the directory that contains the soteria-c.h
*)

no_ignore_parse_failures : bool;
(*
Files that cannot be parsed correctly are ignored by default, this flag deactivates that behaviour.
*)

no_ignore_duplicate_symbols : bool;
(*
Programs that contain duplicate symbols are ignored by default, this flag deactivates that behaviour.
*)

parse_only : bool;
(*
Only parse and link the C program, do not perform analysis
*)

no_c23 : bool;
(*
Disable C23 support (even if the underlying Cerberus library supports it).
*)

dump_summaries_file : string option;
(*
Dump the generated summaries to a file
*)

show_manifest_summaries : bool;
(*
Print a corresponding manifest summary after the bug report if a bug is found
*)

alloc_cannot_fail : bool;
(*
Assume allocations cannot fail
*)

use_cerb_headers : bool;
(*
Use the Cerberus-provided standard headers instead of the system headers.
*)

cbmc_compat : bool;
(*
Enable support for a subset of the __CPROVER_ API.
*)

testcomp_compat : bool;
(*
Enable support for a subset of the testcomp API (e.g., __VERIFIER_nondet_
*)

ignore_ub : bool;
(*
Ignores undefined behaviour branches (mostly for Test-Comp's weird requirements). Branches reaching UB will be dismissed. Only has effect in symbolic testing mode.
*)

havoc_undefined_funs : bool;
(*
Assume that all undefined functions can return any value. Warning: this can lead to unsoundnesses in analyses.
*)

print_states : bool;
(*
Print final program states after whole-program symbolic testing
*)

write_parsed_db : string option;
(*
When using a compilation database, write a filtered version containing only successfully parsed files to the specified path
*)

dump_report : string option;
(*
Write a JSON report of all diagnostics (bugs and errors) to the specified file
*)

}

val make : ?auto_include_path:string -> ?no_ignore_parse_failures:bool -> ?no_ignore_duplicate_symbols:bool -> ?parse_only:bool -> ?no_c23:bool -> ?dump_summaries_file:string -> ?show_manifest_summaries:bool -> ?alloc_cannot_fail:bool -> ?use_cerb_headers:bool -> ?cbmc_compat:bool -> ?testcomp_compat:bool -> ?ignore_ub:bool -> ?havoc_undefined_funs:bool -> ?print_states:bool -> ?write_parsed_db:string -> ?dump_report:string -> unit -> t

val cmdliner_term : unit -> t Cmdliner.Term.t

type mode =

| Compositional

| Whole_program

type Stdlib.Effect.t +=

| GetConfig : t Stdlib.Effect.t

type Stdlib.Effect.t +=

| GetMode : mode Stdlib.Effect.t

val default : t

val current : unit -> t

val current_mode : unit -> mode

val test_comp_warning : Soteria.Soteria_std.String.Interned.t

val with_config : config:t -> mode:mode -> (unit -> 'a) -> 'a