Module Stats.As_ctx

Module for manipulating statistics as a context using algebraic effects (that are hidden). All calls to any function in this module should happen only inside a function wrapped with with_stats, ensuring that the statistics are properly passed around.

val with_stats : unit -> (unit -> 'a) -> 'a with_stats

with_stats () f runs function f and handles effects raised by the functions of this module such as add_exec_time, and returns a record containing the result of executing f together with the obtained statistics.

val with_stats_ignored : unit -> (unit -> 'a) -> 'a

with_stats_ignored () f runs function f and handles effects raised by the functions of this module, but ignores their effect. This is to be used when the user does not wish to pay the (minor) performance cost of stats bookkeeping.

val add_exec_time : float -> unit

Adds the given execution time to the stats record. Should be handled by Soteria itself.

val add_time_of : (unit -> 'a) -> 'a

Measures time taken by the given function and adds it to the statistics in the environment.

val add_sat_time : float -> unit

Adds the time taken to answer sat to the stats record. Should be handled by Soteria itself.

val add_sat_time_of : (unit -> 'a) -> 'a

Measures time taken by the given function and adds it to the SAT time statistics in the environment.

val add_branches : int -> unit

Adds branches to the count of statistics. Should be handled by Soteria itself

val add_steps : int -> unit

Adds steps to the count of statistics. Handled by Soteria when users call Symex.S.consume_fuel_steps

val add_unexplored_branches : int -> unit

Adds unexplored branches due to fuel exhaustion to the count of statistics. Handled by Soteria when users call Symex.S.consume_fuel_steps or when branching

val add_sat_checks : int -> unit

Adds the number of times the SAT solver was called. Handled by Soteria.

val add_sat_unknowns : int -> unit

Adds the number of times the SAT solver returned unknowns to the count of statistics. Handled by Soteria.

val push_missing_without_fix : string -> unit

Push a reason for not being able to fix a Missing error. Handled by Soteria when users call Soteria.Symex.S.Result.miss_no_fix

val push_give_up_reason : loc:Range.t -> string -> unit

Push a reason for giving up symex, with the location at which it was reached. Handled by Soteria when users call Soteria.Symex.S.give_up