Module SM.Value

type +'a t

The type of symbolic values, parameterized by their type. For example, sbool t represents a symbolic boolean.

type +'a ty

Type of values.

type sbool

The type for booleans.

val not : sbool t -> sbool t

not b returns the logical negation of boolean b.

val ppa : Stdlib.Format.formatter -> 'a t -> unit

val iter_vars : 'a t -> 'b ty Soteria.Symex.Var.iter_vars

iter_vars v f iterates over all variables in value v, calling f (var, ty) for each variable with its type.

val subst : (Soteria.Symex.Var.t -> Soteria.Symex.Var.t) -> 'a t -> 'a t

subst f v substitutes variables in v using function f. Each variable x is replaced with f x.

val mk_var : Soteria.Symex.Var.t -> 'a ty -> 'a t

mk_var var ty creates a symbolic variable with name var and type ty.

val to_bool : 'a t -> bool option

to_bool v returns Some b if v is a concrete boolean b, or None if v is symbolic.

val of_bool : bool -> sbool t

of_bool b creates a concrete boolean value from b.