Parameter Tree_block.MemVal

module S_bool : Soteria.Data.S_bool.S(CSYMEX).S

module S_bounded_int : Soteria.Data.S_int.S(CSYMEX)(S_bool).Bounded_S

type t

type sint := S_bounded_int.t CSYMEX.Value.t

val pp : Stdlib.Format.formatter -> t -> unit

val merge : left:t -> right:t -> t

Merges two children node into a single node; returns the merged node. Note children of the node are always preserved too, for further accesses.

val split : at:sint -> t -> ((t, sint) Soteria__Sym_states__Tree_block.Split_tree.t * (t, sint) Soteria__Sym_states__Tree_block.Split_tree.t) CSYMEX.t

split ~at node Splits node at at, which is the relative offset within the node. Returns the left and right split trees, which themselves may contain further splits.

at is guaranteed to be in the range [1, size(node)), i.e. strictly within the node.

type serialized

val pp_serialized : Stdlib.Format.formatter -> serialized -> unit

val show_serialized : serialized -> string

val subst_serialized : (Soteria.Symex.Var.t -> Soteria.Symex.Var.t) -> serialized -> serialized

val iter_vars_serialized : serialized -> ((Soteria.Symex.Var.t * 'b CSYMEX.Value.ty) -> unit) -> unit

val serialize : t -> serialized Stdlib.Seq.t option

Serialize this memory value; either returns Some serialized, or None to signal the children must instead be serialized.

val consume : serialized -> (t, sint) Soteria__Sym_states__Tree_block.tree -> ((t, sint) Soteria__Sym_states__Tree_block.tree, 'err, serialized) CSYMEX.Result.t

Extract the given serialized predicate from the tree; this may result in an empty (NotOwned Totally) tree, or may only modify part of the tree if the predicate only represents part of this tree's state. A Missing may be raised if part of the state is missing for the consumption to succeed.

The input tree corresponds to the subtree relevant to the predicate's offset and length, meaning t.node is the node covering the whole predicate's range.

val produce : serialized -> (t, sint) Soteria__Sym_states__Tree_block.tree -> (t, sint) Soteria__Sym_states__Tree_block.tree CSYMEX.t

Add the given serialized predicate onto the given tree; the input tree is not necessarily empty (NotOwned Totally), and if the predicate overlaps the production may vanish.

The input tree corresponds to the subtree relevant to the predicate's offset and length, meaning t.node is the node covering the whole predicate's range.

val assert_exclusively_owned : t -> (unit, 'err, serialized) CSYMEX.Result.t

Returns ok if this memory value is exclusively owned, ie no additional state can be composed with it; in other words, calling produce on a tree with this node must always vanish. Otherwise this should raise a miss with the fixes needed for this to become exclusively owned.