Module Soteria_c_lib.Ctree_block

module Ctype = Cerb_frontend.Ctype

module MemVal : sig ... end

include sig ... end

type nonrec sint = MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t

type nonrec (!'a, !'sint) tree = {

node : 'a Soteria__Sym_states__Tree_block.node;

range : 'sint * 'sint;

children : (('a, 'sint) Soteria__Sym_states__Tree_block.tree * ('a, 'sint) Soteria__Sym_states__Tree_block.tree) option;

}

module Range : sig ... end

module Split_tree : sig ... end

A Split_tree is a simplified representation of a tree, that has no offset. It however indicates, on Nodes, at what offset the split occurs, relative to that node's start.

module Node : sig ... end

module Tree : sig ... end

type t = {

root : Tree.t;

bound : sint option;

}

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

module SM : sig ... end

val is_empty : t -> bool

val pp_pretty : Stdlib.Format.formatter -> t -> unit

type serialized =

| MemVal of {

offset : sint;

len : sint;

v : MemVal.serialized;

}

| Bound of sint

val pp_serialized : Ppx_deriving_runtime.Format.formatter -> serialized -> Ppx_deriving_runtime.unit

val show_serialized : serialized -> Ppx_deriving_runtime.string

val lift_miss : offset:sint -> len:sint -> (Cerb_location.t -> (('a, 'b, MemVal.serialized) Soteria.Symex.Compo_res.t * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t) -> Cerb_location.t -> (('a, 'b, serialized list) Soteria.Symex.Compo_res.t * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t

val iter_values_serialized : serialized list -> (MemVal.serialized -> unit) -> unit

val of_opt : ?mk_fixes: (unit -> Cerb_location.t -> ('a list * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t) -> 'b option -> Cerb_location.t -> (('b, 'c, 'a) Soteria.Symex.Compo_res.t * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t

val to_opt : t -> t option

val with_bound_check : ?mk_fixes: (unit -> Cerb_location.t -> (serialized list list * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t) -> sint -> (Tree.t -> Cerb_location.t -> (('a * Tree.t, [> `OutOfBounds ] as 'b, serialized list) Soteria.Symex.Compo_res.t * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t) -> ('a, 'b, serialized list) SM.Result.t

val assert_exclusively_owned : t option -> Cerb_location.t -> ((unit, 'a, serialized list) Soteria.Symex.Compo_res.t * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t

val get_raw_tree_owned : MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t -> MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t -> ((MemVal.t, MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t) Soteria__Sym_states__Tree_block.tree, [> `OutOfBounds ], serialized list) SM.Result.t

val put_raw_tree : MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t -> Tree.t -> (unit, [> `OutOfBounds ], serialized list) SM.Result.t

val subst_serialized : (Soteria.Symex.Var.t -> Soteria.Symex.Var.t) -> serialized -> serialized

val iter_vars_serialized : serialized -> ((Soteria.Symex.Var.t * 'a Soteria.Bv_values.Bv_solver.Z3_solver.Value.ty) -> unit) -> unit

val serialize : t -> serialized list

val consume_bound : MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t -> t option -> Cerb_location.t -> ((t option, 'a, 'b) Soteria.Symex.Compo_res.t * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t

val produce_bound : sint -> unit SM.t

val produce_mem_val : sint -> MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t -> MemVal.serialized -> unit SM.t

val consume_mem_val : sint -> sint -> MemVal.serialized -> t option -> Cerb_location.t -> ((t option, 'a, serialized list) Soteria.Symex.Compo_res.t * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t

val produce : serialized -> unit SM.t

val log_fixes : serialized list list -> serialized list list

val range_of_low_and_type : MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t -> Soteria_c_lib.Layout.CF.Ctype.ctype -> Cerb_location.t -> ((MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t * MemVal.S_bounded_int.t Soteria.Bv_values.Bv_solver.Z3_solver.Value.t) * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t

val sval_leaf : range:('a * 'a) -> value:Soteria_c_lib.Typed.T.cval Typed.t -> ty:Ctype.ctype -> (MemVal.t, 'a) Soteria__Sym_states__Tree_block.tree

val zeros : range:('a * 'a) -> (MemVal.t, 'a) Soteria__Sym_states__Tree_block.tree

val uninit : range:('a * 'a) -> (MemVal.t, 'a) Soteria__Sym_states__Tree_block.tree

val mk_fix_typed : sint -> Soteria_c_lib.Layout.CF.Ctype.ctype -> unit -> Cerb_location.t -> (serialized list list * Cerb_location.t) Soteria__Symex.Make(Soteria.Bv_values.Bv_solver.Z3_solver).t

val mk_fix_any : sint -> sint -> unit -> serialized list list

val mk_fix_any_s : sint -> sint -> unit -> serialized list list Csymex.t

val decode : ty:Ctype.ctype -> ofs:sint -> MemVal.t MemVal.TB.node -> (Soteria_c_lib.Typed.T.cval Typed.t, [> `UninitializedMemoryAccess ], serialized list) Csymex.Result.t

val load : Soteria_c_lib.Typed.T.sint Typed.t -> Ctype.ctype -> (Soteria_c_lib.Typed.T.cval Typed.t, [> `OutOfBounds | `UninitializedMemoryAccess ], serialized list) SM.Result.t

val store : [ `NonZero | `Zero ] Typed.t -> Ctype.ctype -> Soteria_c_lib.Typed.T.cval Typed.t -> (unit, [> `OutOfBounds ], serialized list) SM.Result.t

val zero_range : Soteria_c_lib.Typed.T.sint Typed.t -> Soteria_c_lib.Typed.T.sint Typed.t -> (unit, [> `OutOfBounds ], serialized list) SM.Result.t

val deinit : Soteria_c_lib.Typed.T.sint Typed.t -> Soteria_c_lib.Typed.T.sint Typed.t -> (unit, [> `OutOfBounds ], serialized list) SM.Result.t

val alloc : zeroed:bool -> sint -> t