Module type S_bounded_int.Bool_

val v_true : [> sbool ] Soteria.Bv_values.Typed.t

val v_false : [> sbool ] Soteria.Bv_values.Typed.t

val of_bool : bool -> [> sbool ] Soteria.Bv_values.Typed.t

val to_bool : 'a Soteria.Bv_values.Typed.t -> bool option

val and_ : [< sbool ] Soteria.Bv_values.Typed.t -> [< sbool ] Soteria.Bv_values.Typed.t -> [> sbool ] Soteria.Bv_values.Typed.t

val and_lazy : [< sbool ] Soteria.Bv_values.Typed.t -> (unit -> [< sbool ] Soteria.Bv_values.Typed.t) -> [> sbool ] Soteria.Bv_values.Typed.t

Similar to and_, but the rhs is only evaluated if the lhs is not the concrete false. In other words, this is a short-circuiting and. Avoids some errors, like a division by zero in 0 != x && n / x when x is 0.

val conj : [< sbool ] Soteria.Bv_values.Typed.t list -> [> sbool ] Soteria.Bv_values.Typed.t

val split_ands : [< sbool ] Soteria.Bv_values.Typed.t -> ([> sbool ] Soteria.Bv_values.Typed.t -> unit) -> unit

val or_ : [< sbool ] Soteria.Bv_values.Typed.t -> [< sbool ] Soteria.Bv_values.Typed.t -> [> sbool ] Soteria.Bv_values.Typed.t

val or_lazy : [< sbool ] Soteria.Bv_values.Typed.t -> (unit -> [< sbool ] Soteria.Bv_values.Typed.t) -> [> sbool ] Soteria.Bv_values.Typed.t

Similar to or_, but the rhs is only evaluated if the lhs is not the concrete true. In other words, this is a short-circuiting or. Avoids some errors, like a division by zero in 0 == x || n / x when x is 0.

val not : [< sbool ] Soteria.Bv_values.Typed.t -> [> sbool ] Soteria.Bv_values.Typed.t

val distinct : 'a Soteria.Bv_values.Typed.t list -> [> sbool ] Soteria.Bv_values.Typed.t

val ite : [< sbool ] Soteria.Bv_values.Typed.t -> 'a Soteria.Bv_values.Typed.t -> 'a Soteria.Bv_values.Typed.t -> 'a Soteria.Bv_values.Typed.t