Module Ctree_block.Split_tree

A Split_tree is a simplified representation of a tree, that has no offset. It however indicates, on Nodes, at what offset the split occurs, relative to that node's start.

type ('a, 'sint) t =

| Leaf of 'a

| Node of ('a, 'sint) t * 'sint * ('a, 'sint) t

val pp : 'a 'sint. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'sint -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> ('a, 'sint) t -> Ppx_deriving_runtime.unit

val show : 'a 'sint. (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'sint -> Ppx_deriving_runtime.unit) -> ('a, 'sint) t -> Ppx_deriving_runtime.string

val map : ('a -> 'b) -> ('a, 'c) t -> ('b, 'c) t